+++ src/unexploitable_search/model_organisms/train.py
@@ -3,65 +3,22 @@ import wandb
 from datasets import load_dataset  # type: ignore
 from hydra.utils import instantiate
 from omegaconf import DictConfig, OmegaConf
-import asyncio
-from typing import Any, Callable
 
-# try:
-#    from unsloth import FastLanguageModel  # type: ignore
-#
-#    UNSLOTH_AVAILABLE = True
-# except ImportError:
-#    FastLanguageModel = None
-#    UNSLOTH_AVAILABLE = False
 from peft import get_peft_model
 from trl.trainer.grpo_trainer import GRPOTrainer
 from trl.trainer.grpo_config import GRPOConfig
 from transformers import AutoModelForCausalLM
-from unexploitable_search.quest import Quest
 from unexploitable_search.config.core import PROJECT_ROOT
-# from trl.trainer.grpo_trainer import RewardFunc
-
-
-def rename_to_prompt(d):  # type: ignore
-    cols = d.column_names
-    if "prompt" in cols:
-        return d
-    if "question" in cols:
-        return d.rename_column("question", "prompt")
-    if "input" in cols:
-        return d.rename_column("input", "prompt")
-    raise KeyError("No 'question' or 'input' column to rename to 'prompt'.")
-
-
-def make_side_quest_reward(side_quest: Quest) -> Callable:  # type: ignore
-    def reward(
-        prompts: list[list[dict[str, str]]],
-        completions: list[list[dict[str, str]]],
-        **kwargs: Any,
-    ) -> list[float]:
-        texts = [c[0]["content"] for c in completions]
-
-        async def run():
-            results = await asyncio.gather(*(side_quest.evaluate(t) for t in texts))
-            return [float(s) for s, _ in results]
-
-        return asyncio.run(run())
-
-    return reward
+from unexploitable_search.model_organisms.train_utils import (
+    make_quest_grpo_reward,
+    rename_to_prompt,
+)
 
 
 @hydra.main(
     version_base=None, config_path=str(PROJECT_ROOT / "conf"), config_name="config"
 )
 def main(cfg: DictConfig):
-    # if not UNSLOTH_AVAILABLE:
-    #     raise ImportError(
-    #         "unsloth is not available on this platform; run on linux with cuda."
-    #     )
-
-    # assert FastLanguageModel is not None, "huh, the last error should have raised"
-
-    # model, tokenizer = FastLanguageModel.from_pretrained(cfg.model.hf_path)
     model = AutoModelForCausalLM.from_pretrained(cfg.model.hf_path)
     lora_cfg = instantiate(cfg.adapter)
     model = get_peft_model(model, lora_cfg)
@@ -69,7 +26,7 @@ def main(cfg: DictConfig):
     wandb.init(
         project=cfg.project,
         name=cfg.run_name,
-        config=OmegaConf.to_container(cfg, resolve=True),  # type: ignore # TODO maybe don't
+        config=OmegaConf.to_container(cfg, resolve=True),  # type: ignore
     )
 
     setting = instantiate(cfg.setting)
@@ -87,10 +44,9 @@ def main(cfg: DictConfig):
 
     trainer = GRPOTrainer(
         model=model,  # type: ignore
-        reward_funcs=make_side_quest_reward(side_quest),
+        reward_funcs=make_quest_grpo_reward(side_quest),
         args=args,
         train_dataset=dataset,  # type: ignore
-        # processor=tokenizer,
     )
 
     output_dir = cfg.train.output_dir if hasattr(cfg.train, "output_dir") else "outputs"
