+++ src/unexploitable_search/model_organisms/train_utils.py
@@ -1,0 +1,41 @@
+from typing import Callable, Any
+import asyncio
+from unexploitable_search.quest import Quest
+
+
+def make_quest_grpo_reward(quest: Quest) -> Callable:  # type: ignore
+    """
+    The Quest.evaluate method requires only a completion, so cannot be used
+    out-of-the-box as a reward for GRPO, because it expects a prompt and completion
+    (even though the prompt is unnecessary for side quest rewards).
+    """
+
+    def reward(
+        prompts: list[list[dict[str, str]]],
+        completions: list[list[dict[str, str]]],
+        **kwargs: Any,
+    ) -> list[float]:
+        texts = [c[0]["content"] for c in completions]
+
+        async def run():
+            results = await asyncio.gather(*(quest.evaluate(t) for t in texts))
+            return [float(s) for s, _ in results]
+
+        return asyncio.run(run())
+
+    return reward
+
+
+def rename_to_prompt(d):  # type: ignore
+    """
+    GRPO expects a dataset column named "prompt".
+    This function renames the "question" or "input" column to "prompt".
+    """
+    cols = d.column_names
+    if "prompt" in cols:
+        return d
+    if "question" in cols:
+        return d.rename_column("question", "prompt")
+    if "input" in cols:
+        return d.rename_column("input", "prompt")
+    raise KeyError("No 'question' or 'input' column to rename to 'prompt'.")
